UI-28:  Changed unit test as a result of the creditor/custom balance changes made in dependent projects.

+++ src/test/java/org/rmt2/soap/accounting/subsidiary/CustomerQuerySoapRequestBuilderTest.java
@@ -68,7 +68,7 @@ public class CustomerQuerySoapRequestBuilderTest {
                 .withAccountNo("ACCT-NO-8888")
                 .withCreditLimit(1234.55)
                 .withAcctDescription("ACCOUNT DESCRIPTION")
-                .withBalance(50000)
+                .withBalance((double) 50000)
                 .withActive(1).build();
                 
         CustomerCriteriaType criteria = fact.createCustomerCriteriaType();
